backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pydantic
python-multipart
reportlab
python-docx
python-pptx
sqlalchemy
python-jose[cryptography]
passlib[argon2]
argon2-cffi
email-validator
numpy>=1.24.0
slowapi
requests
redis
alembic
cryptography
celery
slowapi
matplotlib
pytest
pytest-asyncio
httpx
pytest-benchmark
scikit-learn
groq
alpha_vantage
python-dotenv
psycopg2-binary
uvicorn[standard]
